Header Component Fix / 960px
This commit fixes the issues that came up when using  the classes .container.small with a max-width of 960px.

The original header component code was only designed for the div with a class of container that wraps around the body and spans full screen width.

Check demo_header.html to see both cases.