Renamed index.html & Logo/Brand
I renamed index.html to index_dev.html (as in "deverloper version") and started a home page by saving navExample.html as index.html.

Created a logo based on text, styling, and Font Awesome icon.

This doesn't have to be the final verdict for the name/logo but I think it should suffice. I'm open to suggestions.